# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atelier API
  version: "0.1.0"
  description: |
    REST + SSE API for Atelier (canvas-ai) -- an AI-powered UI design generation platform.
    Supports project management, screen generation (including redesign-from-URL and
    image-to-code), component library CRUD, analytics, and feedback collection.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Optional Bearer token set via ATELIER_API_KEY environment variable.
        When the server has no key configured, all requests are allowed.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
paths:
...
  /api/analytics/stats:
    get:
      tags: [Analytics]
      summary: Get analytics storage stats
      security: []
      responses:
        "200":
          description: Analytics stats with free-tier usage info
          content:
            application/json:
              schema:
                type: object
                properties:
                  estimatedSizeMB:
                    type: number
                  freeTierLimitMB:
                    type: number
                    example: 512
                  usagePercent:
                    type: number
                additionalProperties: true
        "503":
          description: Analytics not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/feedback:
    post:
      tags: [Analytics]
      summary: Log user feedback on a generated screen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, screenId, action]
              properties:
                projectId:
                  type: string
                screenId:
                  type: string
                action:
                  type: string
                  enum: [kept, deleted, edited, exported, variant_created]
                generationId:
                  type: string
                editDelta:
                  type: string
                rating:
                  type: number
                  nullable: true
                timeToAction:
                  type: number
                html:
                  type: string
                  description: >
                    Full HTML of the screen. When action is exported, edited, or
                    variant_created, sections are extracted and saved as templates.
      responses:
        "200":
          description: Feedback recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          description: Missing required field or invalid action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/projects/{pid}/design-system/extract:
    post:
      tags: [Analytics]
      summary: Extract design system tokens from a URL
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
      responses:
        "200":
          description: Extracted design system object
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Invalid URL or missing field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: SSRF - URL points to private network
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/assets/{key}:
    get:
      tags: [Analytics]
      summary: Serve a stored asset file (image)
      security: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Asset filename (alphanumeric, dots, dashes, underscores)
      responses:
        "200":
          description: Asset binary data
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          description: Invalid asset path (path traversal attempt)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"